move processor classes for simplification
