+++ app/processors/geo_concerns/processors/raster.rb
@@ -1,10 +1,25 @@
 module GeoConcerns
   module Processors
     module Raster
-      extend ActiveSupport::Autoload
+      class Processor < Hydra::Derivatives::Processors::Processor
+        def process
+          raster_processor_class.new(source_path,
+                                     directives,
+                                     output_file_service: output_file_service).process
+        end
 
-      eager_autoload do
-        autoload :Processor
+        # Returns a raster processor class based on mime type passed in the directives object.
+        # @return raster processing class
+        def raster_processor_class
+          case directives.fetch(:input_format)
+          when 'text/plain; gdal-format=USGSDEM'
+            GeoConcerns::Processors::Raster::Dem
+          when 'application/octet-stream; gdal-format=AIG'
+            GeoConcerns::Processors::Raster::Aig
+          else
+            GeoConcerns::Processors::Raster::Base
+          end
+        end
       end
     end
   end
